// src/types.ts
export type ExpenseAmount = {
  amount: number;
  currencyCode: string;
};

export type Expense = {
  id: number;
  amount: ExpenseAmount;
  category: string;
  description: string;
  date: string;
  blockId: number | null;
  paidByUserId: number;
  paidByUser: { type: string; id: number };
  splitWith: { type: string; users: Array<{ type: string; id: number }> };
};